Move clean_cached_assets_randomly to libapi
